# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Youtube Video API",
        "description": "This is a Youtube Video API that allow users to retrieve generated data about youtube video(s) as a JSON format.",
        "contact": {
            "email": "n00201327@iadt.ie"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/youtubeVideos": {
            "get": {
                "tags": [
                    "Youtube Videos"
                ],
                "summary": "Display all videos",
                "description": "Displays all the youtube videos with its relationship.",
                "parameters": [
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "orderBy allows data to be re-arrange in the order of title, created_at, views, likes, and dislikes.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "random",
                        "in": "query",
                        "description": "retrieve data and pluck it in random order.",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "likes",
                        "in": "query",
                        "description": "retrieve data by filtering the videos that have the amount of likes less than the specified number.",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "dislikes",
                        "in": "query",
                        "description": "retrieve data by filtering the videos that have the amount of dislikes less than the specified number.",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "description": "retrieve data by filtering the videos that are made in the specified year.",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "hasVideos",
                        "in": "query",
                        "description": "retrieve data by filtering the channels that have videos.",
                        "schema": {
                            "type": "enum",
                            "enum": [
                                "yes",
                                "no"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation, Returns a list of Videos in JSON format",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/youtube_video"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Youtube Videos"
                ],
                "summary": "Create a new youtube video",
                "description": "update and store the updated data to a specific youtube video in the database",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "duration",
                                    "likes",
                                    "dislikes",
                                    "views"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Sample Title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "A long description about this great book"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "60"
                                    },
                                    "likes": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "200"
                                    },
                                    "dislikes": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "300"
                                    },
                                    "views": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/youtube_video"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/youtubeVideos/{id}": {
            "get": {
                "tags": [
                    "Youtube Videos"
                ],
                "summary": "Gets a youtube video by its ID",
                "description": "Retreive a youtube video specified by the ID parameter and displays it in a JSON format.",
                "operationId": "5b85c2afa115dc9c8692ed462f88004a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "video id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/youtube_video"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Youtube Videos"
                ],
                "summary": "Update a youtube video by id",
                "description": "update and store the updated data to a specific youtube video in the database",
                "operationId": "put",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "duration",
                                    "likes",
                                    "dislikes",
                                    "views"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Sample Title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "A long description about this great book"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "60"
                                    },
                                    "likes": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "200"
                                    },
                                    "dislikes": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "300"
                                    },
                                    "views": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Video updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/youtube_video"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid video ID"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Youtube Videos"
                ],
                "summary": "Delete a youtube video by its ID",
                "description": "Delete a youtube video specified by the ID parameter.",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a Video",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/successful_delete"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Video not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/channel/videos": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get video(s) made by your channel",
                "description": "retrieve video(s), given it existed, from the token bearer.",
                "operationId": "37f4cf59747de604efbe76fed2d1b110",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/youtube_video"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorised"
                    },
                    "404": {
                        "description": "Comment Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/youtubeVideos/{id}/comments": {
            "get": {
                "tags": [
                    "Youtube Videos"
                ],
                "summary": "Gets comments by its youtube video ID",
                "description": "Retreive comments from a youtube video specified by the ID parameter and displays it in a JSON format.",
                "operationId": "096a9f120a1259a3a23b85893bf3783f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Video ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "successful_delete": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The data has been successfully deleted"
                    },
                    "status": {
                        "type": "integer",
                        "example": "202"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "0": []
        }
    },
...
}